requirements.txt
Django==1.4
MySQL-python==1.2.3
distribute==0.6.27
dj-database-url==0.2.1
django-tastypie==0.9.11
gunicorn==0.14.5
mimeparse==0.1.3
psycopg2==2.4.5
python-dateutil==1.5
wsgiref==0.1.2
